Fully transitioned into new file structure
-- React frontend sending requests to backend and receiving responses
through rudimentary forms right now, but all routes are functional in a
rudimentary way now and backend is storing images in temp folder for
processing, then posting image ascii art to an array of structs that remains in
state. The process is still relatively quick and I'm very proud of it.

-- I need to hook a database to this soon
-- I need to implement deletion of tempfiles soon as well
-- I need to write more form validation for image submission, but I
think that might be better to do from the frontend with a library like
yup
-- Overall great breakthrough progress, but much work remains to be
done.